+++ internal/command/format/diagnostic_test.go
@@ -267,10 +267,105 @@ func TestDiagnostic(t *testing.T) {
 [red]╵[reset]
 `,
 		},
+		"test number assertion difference": {
+			&hcl.Diagnostic{
+				Severity: hcl.DiagError,
+				Summary:  "Bad testing",
+				Detail:   "Number testing went wrong.",
+				Expression: &hclsyntax.BinaryOpExpr{
+					LHS: &hclsyntax.LiteralValueExpr{
+						Val: cty.StringVal("3"),
+					},
+					RHS: &hclsyntax.LiteralValueExpr{
+						Val: cty.StringVal("5"),
+					},
+					Op: hclsyntax.OpEqual,
+				},
+				Subject: &hcl.Range{
+					Filename: "test.tf",
+					Start:    hcl.Pos{Line: 1, Column: 6, Byte: 5},
+					End:      hcl.Pos{Line: 1, Column: 12, Byte: 11},
+				},
+				EvalContext: &hcl.EvalContext{},
+			},
+			`[red]╷[reset]
+[red]│[reset] [bold][red]Error: [reset][bold]Bad testing[reset]
+[red]│[reset]
+[red]│[reset]   on test.tf line 1:
+[red]│[reset]    1: test [underline]source[reset] code
+[red]│[reset]     [dark_gray]├────────────────[reset]
+[red]│[reset]     [dark_gray]│[reset] [bold]Diff: [reset]
+[red]│[reset]     [dark_gray]│[reset]     "3" [yellow]->[reset] "5"
+[red]│[reset]
+[red]│[reset] Number testing went wrong.
+[red]╵[reset]
+`,
+		},
+		"test object assertion difference": {
+			&hcl.Diagnostic{
+				Severity: hcl.DiagError,
+				Summary:  "Bad testing",
+				Detail:   "Test object assertion!",
+				Expression: &hclsyntax.BinaryOpExpr{
+					LHS: &hclsyntax.ScopeTraversalExpr{
+						Traversal: hcl.Traversal{
+							hcl.TraverseRoot{Name: "var"},
+							hcl.TraverseAttr{Name: "json_headers"},
+						},
+					},
+					RHS: &hclsyntax.LiteralValueExpr{
+						Val: cty.ObjectVal(map[string]cty.Value{
+							"Test-Header-1": cty.StringVal("foo"),
+							"Test-Header-2": cty.StringVal("bar"),
+						}),
+					},
+					Op: hclsyntax.OpEqual,
+				},
+				Subject: &hcl.Range{
+					Filename: "json_encode.tf",
+					Start:    hcl.Pos{Line: 1, Column: 12, Byte: 12},
+					End:      hcl.Pos{Line: 4, Column: 20, Byte: 150},
+				},
+				EvalContext: &hcl.EvalContext{
+					Variables: map[string]cty.Value{
+						"var": cty.ObjectVal(map[string]cty.Value{
+							"json_headers": cty.ObjectVal(map[string]cty.Value{
+								"Test-Header-1": cty.StringVal("foo"),
+								"Test-Header-2": cty.StringVal("foo"),
+							}),
+						}),
+					},
+				},
+			},
+			`[red]╷[reset]
+[red]│[reset] [bold][red]Error: [reset][bold]Bad testing[reset]
+[red]│[reset]
+[red]│[reset]   on json_encode.tf line 1:
+[red]│[reset]    1: condition = [underline]jsonencode(var.json_headers) == jsonencode([
+[red]│[reset]    2: 			"Test-Header-1: foo",
+[red]│[reset]    3: 			"Test-Header-2: bar"
+[red]│[reset]    4: 		])[reset]
+[red]│[reset]     [dark_gray]├────────────────[reset]
+[red]│[reset]     [dark_gray]│[reset] [bold]var.json_headers[reset] is object with 2 attributes
+[red]│[reset]     [dark_gray]├────────────────[reset]
+[red]│[reset]     [dark_gray]│[reset] [bold]Diff: [reset]
+[red]│[reset]     [dark_gray]│[reset]     {
+[red]│[reset]     [dark_gray]│[reset]       [yellow]~[reset] Test-Header-2 = "foo" [yellow]->[reset] "bar"
+[red]│[reset]     [dark_gray]│[reset]         [dark_gray]# (1 unchanged attribute hidden)[reset]
+[red]│[reset]     [dark_gray]│[reset]     }
+[red]│[reset]
+[red]│[reset] Test object assertion!
+[red]╵[reset]
+`,
+		},
 	}
 
 	sources := map[string]*hcl.File{
 		"test.tf": {Bytes: []byte(`test source code`)},
+		"json_encode.tf": {Bytes: []byte(`condition = jsonencode(var.json_headers) == jsonencode([
+			"Test-Header-1: foo",
+			"Test-Header-2: bar"
+		])`)},
 	}
 
 	// This empty Colorize just passes through all of the formatting codes
@@ -284,8 +379,8 @@ func TestDiagnostic(t *testing.T) {
 			diag := diags[0]
 			got := strings.TrimSpace(Diagnostic(diag, sources, colorize, 40))
 			want := strings.TrimSpace(test.Want)
-			if got != want {
-				t.Errorf("wrong result\ngot:\n%s\n\nwant:\n%s\n\n", got, want)
+			if diff := cmp.Diff(want, got); diff != "" {
+				t.Errorf("diff:\n%s", diff)
 			}
 		})
 	}
@@ -588,8 +683,8 @@ Whatever shall we do?
 			diag := diags[0]
 			got := strings.TrimSpace(DiagnosticPlain(diag, sources, 40))
 			want := strings.TrimSpace(test.Want)
-			if got != want {
-				t.Errorf("wrong result\ngot:\n%s\n\nwant:\n%s\n\n", got, want)
+			if diff := cmp.Diff(want, got); diff != "" {
+				t.Errorf("diff:\n%s", diff)
 			}
 		})
 	}
@@ -702,8 +797,8 @@ func TestDiagnostic_nonOverlappingHighlightContext(t *testing.T) {
 `
 	output := Diagnostic(diags[0], sources, color, 80)
 
-	if output != expected {
-		t.Fatalf("unexpected output: got:\n%s\nwant\n%s\n", output, expected)
+	if diff := cmp.Diff(output, expected); diff != "" {
+		t.Errorf("diff:\n%s", diff)
 	}
 }
 
@@ -750,8 +845,8 @@ func TestDiagnostic_emptyOverlapHighlightContext(t *testing.T) {
 `
 	output := Diagnostic(diags[0], sources, color, 80)
 
-	if output != expected {
-		t.Fatalf("unexpected output: got:\n%s\nwant\n%s\n", output, expected)
+	if diff := cmp.Diff(output, expected); diff != "" {
+		t.Errorf("diff:\n%s", diff)
 	}
 }
 
@@ -793,8 +888,8 @@ Error: Some error
 `
 	output := DiagnosticPlain(diags[0], sources, 80)
 
-	if output != expected {
-		t.Fatalf("unexpected output: got:\n%s\nwant\n%s\n", output, expected)
+	if diff := cmp.Diff(output, expected); diff != "" {
+		t.Errorf("diff:\n%s", diff)
 	}
 }
 
@@ -826,8 +921,8 @@ func TestDiagnostic_wrapDetailIncludingCommand(t *testing.T) {
 `
 	output := Diagnostic(diags[0], nil, color, 76)
 
-	if output != expected {
-		t.Fatalf("unexpected output: got:\n%s\nwant\n%s\n", output, expected)
+	if diff := cmp.Diff(output, expected); diff != "" {
+		t.Errorf("diff:\n%s", diff)
 	}
 }
 
@@ -854,8 +949,8 @@ eventually make it onto multiple lines. THE END
 `
 	output := DiagnosticPlain(diags[0], nil, 76)
 
-	if output != expected {
-		t.Fatalf("unexpected output: got:\n%s\nwant\n%s\n", output, expected)
+	if diff := cmp.Diff(output, expected); diff != "" {
+		t.Errorf("diff:\n%s", diff)
 	}
 }
 
@@ -905,8 +1000,8 @@ func TestDiagnosticFromJSON_invalid(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			got := strings.TrimSpace(DiagnosticFromJSON(test.Diag, colorize, 40))
 			want := strings.TrimSpace(test.Want)
-			if got != want {
-				t.Errorf("wrong result\ngot:\n%s\n\nwant:\n%s\n\n", got, want)
+			if diff := cmp.Diff(got, want); diff != "" {
+				t.Errorf("wrong result\n%s", diff)
 			}
 		})
 	}