+++ internal/configs/hcl2shim/util.go
@@ -6,8 +6,11 @@
 package hcl2shim
 
 import (
+	"fmt"
+
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
+	"github.com/zclconf/go-cty/cty"
 )
 
 // exprIsNativeQuotedString determines whether the given expression looks like
@@ -81,6 +84,16 @@ func ConvertJSONExpressionToHCL(expr hcl.Expression) (hcl.Expression, hcl.Diagno
 	value, ds := expr.Value(nil)
 	diags = append(diags, ds...)
 	if diags.HasErrors() {
+		return nil, diags
+	}
+
+	if value.Type() != cty.String || value.IsNull() {
+		diags = append(diags, &hcl.Diagnostic{
+			Severity: hcl.DiagError,
+			Summary:  "Expected string expression",
+			Detail:   fmt.Sprintf("This value must be a string, but got %s.", value.Type().FriendlyName()),
+			Subject:  expr.Range().Ptr(),
+		})
 		return nil, diags
 	}